admin_panel/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
aiohttp==3.9.1
python-dotenv>=1.0.1
pydantic>=2.7.4,<3.0.0
python-multipart==0.0.6

